test(post-service): 게시글 댓글 목록 조회 단위 테스트 추가 #46

+++ src/main/java/com/sparta/twitNation/dto/comment/resp/CommentListRespDto.java
@@ -9,7 +9,7 @@ import java.util.List;
 public record CommentListRespDto(
         long elementsCount,
         long currentPage,
-        long nextPage,
+        Long nextPage,
         boolean hasNextPage,
         long pageSize,
         List<CommentRespDto> commentList
@@ -18,7 +18,7 @@ public record CommentListRespDto(
         this(
                 commentPage.getTotalElements(),
                 commentPage.getNumber(),
-                commentPage.hasNext() ? commentPage.getNumber() + 1 : null,
+                commentPage.hasNext() ? (long) commentPage.getNumber() + 1 : null,
                 commentPage.hasNext(),
                 commentPage.getSize(),
                 commentPage.stream()
